requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai==0.28.1
